+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Python: Added SMOVE command ([#1421](https://github.com/aws/glide-for-redis/pull/1421))
 * Python: Added SUNIONSTORE command ([#1423](https://github.com/aws/glide-for-redis/pull/1423))
 * Python: Added SINTER command ([#1434](https://github.com/aws/glide-for-redis/pull/1434))
+* Python: Added SDIFF command ([#1437](https://github.com/aws/glide-for-redis/pull/1437))
 
 
 #### Fixes